requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# AI APIs
google-generativeai==0.8.0

# Export
python-docx==1.1.0
reportlab==4.0.8
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2
